+++ src/app/core/index.ts
@@ -1,1 +1,2 @@
 export * from './guards/auth.guard';
+export * from './services/feedbacks.service';