+++ src/app/layout/layout.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 
 import { SharedModule } from '../shared/shared.module';
 import { AboutComponent, PathNotFoundComponent, LoginComponent } from './components';
@@ -12,6 +13,7 @@ import { AboutComponent, PathNotFoundComponent, LoginComponent } from './compone
   ],
   imports: [
     CommonModule,
+    FormsModule,
     SharedModule
   ]
 })