+++ src/app/products/products-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { ProductListComponent, ProductDetailsComponent } from './components';
-
+import { FeedbacksComponent } from '../layout/components/';
 const routes: Routes = [
   {
     path: 'product-list',
@@ -11,7 +11,15 @@ const routes: Routes = [
   {
     path: 'product/:productID',
     component: ProductDetailsComponent,
-  },
+    // router-outlet is inside ProductDetailsComponent template => should be in children
+    children: [
+      {
+        path: 'feedbacks',
+        component: FeedbacksComponent,
+        outlet: 'feedbacks'
+      }
+    ]
+  }
 ];
 
 @NgModule({