+++ src/app/products/products.module.ts
@@ -1,18 +1,22 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 
 import { ProductsRoutingModule } from './products-routing.module';
 import { SharedModule } from '../shared/shared.module';
 import { ProductComponent, ProductListComponent, ProductDetailsComponent } from './components';
+import { FeedbacksComponent } from '../layout/components';
 
 @NgModule({
   declarations: [
     ProductComponent,
     ProductListComponent,
-    ProductDetailsComponent
+    ProductDetailsComponent,
+    FeedbacksComponent 
   ],
   imports: [
     CommonModule,
+    FormsModule,
     SharedModule,
     ProductsRoutingModule
   ]